Moving the parsing to its own file
